factored out the row-major curve for unit testing; fixing unit tests
